Revert SenML name to messages
In order to keep backward compatibility.
