+++ writers/cassandra/init.go
@@ -3,9 +3,14 @@
 
 package cassandra
 
-import "github.com/gocql/gocql"
+import (
+	"fmt"
 
-const table = `CREATE TABLE IF NOT EXISTS senml (
+	"github.com/gocql/gocql"
+)
+
+const (
+	table = `CREATE TABLE IF NOT EXISTS messages (
         id uuid,
         channel text,
         subtopic text,
@@ -22,6 +27,18 @@ const table = `CREATE TABLE IF NOT EXISTS senml (
     	update_time double,
         PRIMARY KEY (channel, time, id)
 	) WITH CLUSTERING ORDER BY (time DESC)`
+
+	jsonTable = `CREATE TABLE IF NOT EXISTS %s (
+        id uuid,
+        channel text,
+        subtopic text,
+        publisher text,
+        protocol text,
+        created bigint,
+        payload text,
+        PRIMARY KEY (channel, created, id)
+	) WITH CLUSTERING ORDER BY (created DESC)`
+)
 
 // DBConfig contains Cassandra DB specific parameters.
 type DBConfig struct {
@@ -49,6 +66,9 @@ func Connect(cfg DBConfig) (*gocql.Session, error) {
 	}
 
 	if err := session.Query(table).Exec(); err != nil {
+		return nil, err
+	}
+	if err := session.Query(fmt.Sprintf(jsonTable, "json")).Exec(); err != nil {
 		return nil, err
 	}
